fix(numeric): accumulate mean/var/covariance in f64 to avoid f32 cancellation (PMAT-916) (#2205)
Pillar-1 numerical-correctness beat (sklearn/numpy parity).

StandardScaler.fit and PCA.fit accumulated their per-feature mean and their
variance / covariance cross-products in f32. On large-magnitude, ill-conditioned
data (feature values ~1e6 with sub-unit variance) this suffers catastrophic
cancellation: an f32 sum of 4096 values near 1e6 overflows the f32 mantissa, so
the mean is biased and the resulting variance / covariance is pure rounding
noise. numpy and sklearn accumulate reductions in float64 even for float32 input.

Measured RED (n=4096, base 1e6):
  - StandardScaler std: f32 ~30 vs f64 reference 0.494  (~60x wrong)
  - PCA trace(cov):     f32 ~1922 vs f64 reference 0.192 (~10000x wrong)

Fix: accumulate the mean sum and the variance / covariance cross-products in f64
(centering PCA residuals in f64 against the f64 mean), then store the f32 result.
The public f32 API is unchanged.

Contracts:
  - OBLIG-SCALER-F64-ACCUM (preprocessing-normalization-v1.yaml, FALSIFY-PP-007)
  - OBLIG-PCA-F64-ACCUM    (pca-v1.yaml, FALSIFY-PCA-005)

Falsifiers (mutation-verified non-tautological — reverting the mean accumulator
to f32 flips each RED on the ill-conditioned input):
  - falsify_pn_007_standard_scaler_f64_accumulation
  - falsify_pca_005_f64_covariance_accumulation

pv validate + pv lint contracts/ pass. cargo test -p aprender-core --lib: 13997 pass.

+++ crates/aprender-core/src/preprocessing/pca.rs
@@ -290,34 +290,41 @@ impl Transformer for PCA {
             return Err("n_components cannot exceed number of features".into());
         }
 
-        // Compute mean
-        let mut mean = vec![0.0; n_features];
+        // Compute mean.
+        // OBLIG-PCA-F64-ACCUM: accumulate the per-feature sum in f64 (numpy/sklearn use a
+        // float64 accumulator even for float32 input) so the mean of large-magnitude data
+        // does not lose precision. The stored mean stays f32 (public API unchanged).
+        let mut mean = vec![0.0_f32; n_features];
         #[allow(clippy::needless_range_loop)]
         for j in 0..n_features {
-            let mut sum = 0.0;
+            let mut sum = 0.0_f64;
             for i in 0..n_samples {
-                sum += x.get(i, j);
+                sum += f64::from(x.get(i, j));
             }
-            mean[j] = sum / n_samples as f32;
+            mean[j] = (sum / n_samples as f64) as f32;
         }
 
-        // Center the data
-        let mut centered = vec![0.0; n_samples * n_features];
+        // Center the data in f64 against the f64-derived mean so the centered residuals
+        // (which can be tiny relative to the raw values) keep full precision before the
+        // covariance cross-products are formed.
+        let mut centered = vec![0.0_f64; n_samples * n_features];
         for i in 0..n_samples {
             for j in 0..n_features {
-                centered[i * n_features + j] = x.get(i, j) - mean[j];
+                centered[i * n_features + j] = f64::from(x.get(i, j)) - f64::from(mean[j]);
             }
         }
 
-        // Compute covariance matrix: Σ = (X^T X) / (n-1)
-        let mut cov = vec![0.0; n_features * n_features];
+        // Compute covariance matrix: Σ = (X^T X) / (n-1).
+        // OBLIG-PCA-F64-ACCUM: accumulate the cross-products in f64; without this the
+        // covariance of ill-conditioned data collapses to noise (or goes negative) in f32.
+        let mut cov = vec![0.0_f32; n_features * n_features];
         for i in 0..n_features {
             for j in 0..n_features {
-                let mut sum = 0.0;
+                let mut sum = 0.0_f64;
                 for k in 0..n_samples {
                     sum += centered[k * n_features + i] * centered[k * n_features + j];
                 }
-                cov[i * n_features + j] = sum / (n_samples - 1) as f32;
+                cov[i * n_features + j] = (sum / (n_samples - 1) as f64) as f32;
             }
         }
